+++ crates/matrix-sdk/src/room/mod.rs
@@ -176,7 +176,7 @@ pub mod reply;
 pub mod privacy_settings;
 
 #[cfg(feature = "e2e-encryption")]
-mod shared_room_history;
+pub(crate) mod shared_room_history;
 
 /// A struct containing methods that are common for Joined, Invited and Left
 /// Rooms
@@ -375,7 +375,6 @@ impl Room {
     ///
     /// Only invited and left rooms can be joined via this method.
     #[doc(alias = "accept_invitation")]
-    #[instrument(skip_all, fields(room_id = ?self.inner.room_id()))]
     pub async fn join(&self) -> Result<()> {
         let prev_room_state = self.inner.state();
 
@@ -386,39 +385,7 @@ impl Room {
             ))));
         }
 
-        let inviter = if prev_room_state == RoomState::Invited {
-            match self.invite_details().await {
-                Ok(details) => details.inviter,
-                Err(e) => {
-                    warn!("No invite details were found, can't attempt to find a room key bundle to accept: {e:?}");
-                    None
-                }
-            }
-        } else {
-            None
-        };
-
-        #[cfg(feature = "e2e-encryption")]
-        let enable_share_history_on_invite = self.client.inner.enable_share_history_on_invite;
-
-        #[cfg(not(feature = "e2e-encryption"))]
-        let enable_share_history_on_invite = false;
-
-        debug!(
-            ?prev_room_state,
-            inviter=?inviter.as_ref().map(|room_member| room_member.user_id()),
-            enable_share_history_on_invite,
-            "Joining room",
-        );
-
         self.client.join_room_by_id(self.room_id()).await?;
-
-        #[cfg(feature = "e2e-encryption")]
-        if enable_share_history_on_invite {
-            if let Some(inviter) = inviter {
-                shared_room_history::maybe_accept_key_bundle(self, inviter.user_id()).await?;
-            }
-        }
 
         Ok(())
     }