+++ testing/matrix-sdk-integration-testing/src/tests/e2ee/shared_history.rs
@@ -86,10 +86,10 @@ async fn test_history_share_on_invite() -> Result<()> {
         "io.element.msc4268.room_key_bundle"
     );
 
-    let bob_room = bob.get_room(alice_room.room_id()).expect("Bob should have received the invite");
+    bob.get_room(alice_room.room_id()).expect("Bob should have received the invite");
 
-    bob_room
-        .join()
+    let bob_room = bob
+        .join_room_by_id(alice_room.room_id())
         .instrument(bob_span.clone())
         .await
         .expect("Bob should be able to accept the invitation from Alice");